Fix #134 for publishing with mill 0.12
