+++ .mill-version
@@ -1,1 +1,1 @@
-0.12.4
+0.12.5
