+++ plugin/src-mill0.11/io/kipp/mill/ci/release/Discover.scala
@@ -1,6 +1,9 @@
 package io.kipp.mill.ci.release
 
+import mainargs.TokensReader
+import mill.eval.Evaluator
+
 private[release] object Discover {
-  implicit def millEvaluatorTokenReader =
+  implicit def millEvaluatorTokenReader: TokensReader[Evaluator] =
     mill.main.TokenReaders.millEvaluatorTokenReader
 }
