+++ plugin/src-mill0.12/io/kipp/mill/ci/release/Discover.scala
@@ -1,7 +1,9 @@
 package io.kipp.mill.ci.release
 
+import mainargs.TokensReader
+import mill.eval.Evaluator
+
 private[release] object Discover {
-  implicit def millEvaluatorTokenReader
-      : mainargs.TokensReader[mill.eval.Evaluator] =
+  implicit def millEvaluatorTokenReader: TokensReader[Evaluator] =
     mill.main.TokenReaders.millEvaluatorTokenReader
 }
